Remove conversion warnings
Uncomplete previous commit ca2db24915b9e2b260e1e635822f3db741be86a7
where this changeset should have been part of.

+++ src/chatdDb.h
@@ -232,7 +232,7 @@ public:
                 chatd::MsgCommand *msgCmd = new chatd::MsgCommand(opcode, mChat.chatId(), userid, msgid, ts, updated, keyid);
                 Buffer buf;
                 stmt.blobCol(11, buf);
-                msgCmd->setMsg(buf.buf(), static_cast<size_t>(buf.dataSize()));
+                msgCmd->setMsg(buf.buf(), buf.dataSize());
 
                 queue.back().msgCmd = msgCmd;
             }
@@ -246,7 +246,7 @@ public:
                 chatd::KeyCommand *keyCmd = new chatd::KeyCommand(mChat.chatId(), keyid);
                 Buffer buf;
                 stmt.blobCol(12, buf);
-                keyCmd->setKeyBlobs(buf.buf(), static_cast<uint32_t>(buf.dataSize()));
+                keyCmd->setKeyBlobs(buf.buf(), buf.dataSize());
 
                 queue.back().keyCmd = keyCmd;
             }
